ref: Extract PayloadComparator from _ParityChecker with format_value param
Move _ParityChecker and _qualify from testutils/helpers/serializer_parity.py
to sentry/utils/payload_comparison.py so it can be used by production code.

Changes from the original:
- Renamed _ParityChecker -> PayloadComparator, _qualify -> qualify
- Added format_value parameter (defaults to repr, preserving current behavior)
- Added describe_value() for PII-safe value formatting in production

serializer_parity.py now imports from the new location; all existing callers
of assert_serializer_parity are unaffected.
