+++ src/sentry/testutils/helpers/serializer_parity.py
@@ -1,8 +1,9 @@
 from __future__ import annotations
 
 from collections.abc import Mapping
-from dataclasses import dataclass, field
 from typing import Any
+
+from sentry.utils.payload_comparison import PayloadComparator
 
 
 def assert_serializer_parity(
@@ -50,7 +51,7 @@ def assert_serializer_parity(
     """
     known_diffs = frozenset(known_differences or ())
     unreliable_fields = frozenset(unreliable or ())
-    checker = _ParityChecker()
+    checker = PayloadComparator(format_value=repr)
     checker.compare(old, new, known_diffs, unreliable=unreliable_fields)
 
     assert not checker.mismatches, "Serializer differences:\n" + "\n".join(checker.mismatches)
@@ -61,99 +62,3 @@ def assert_serializer_parity(
         "Unnecessary known_differences (no actual difference found):\n"
         + "\n".join(sorted(unnecessary))
     )
-
-
-def _qualify(prefix: str, name: str) -> str:
-    return f"{prefix}.{name}" if prefix else name
-
-
-@dataclass
-class _ParityChecker:
-    mismatches: list[str] = field(default_factory=list)
-
-    # known_diffs entries confirmed to be actual differences.
-    confirmed: set[str] = field(default_factory=set)
-
-    def _nested_fields(self, field_set: frozenset[str], key: str) -> frozenset[str]:
-        """Extract child paths for *key* from a dot-separated field set.
-
-        E.g. ``_nested_fields({"activities.id", "title"}, "activities")``
-        returns ``{"id"}``.
-        """
-        prefix = key + "."
-        return frozenset(e[len(prefix) :] for e in field_set if e.startswith(prefix))
-
-    def compare(
-        self,
-        old: Mapping[str, Any],
-        new: Mapping[str, Any],
-        known_diffs: frozenset[str],
-        path: str = "",
-        diffs_path: str = "",
-        *,
-        unreliable: frozenset[str] = frozenset(),
-    ) -> None:
-        for key in set(list(old.keys()) + list(new.keys())):
-            if key in known_diffs:
-                full_diffs_key = _qualify(diffs_path, key)
-                if key not in new or key not in old or old[key] != new[key]:
-                    self.confirmed.add(full_diffs_key)
-                continue
-
-            if key in unreliable:
-                full_path = _qualify(path, key)
-                if key not in new:
-                    self.mismatches.append(f"Missing from new: {full_path}")
-                elif key not in old:
-                    self.mismatches.append(f"Extra in new: {full_path}")
-                continue
-
-            full_path = _qualify(path, key)
-
-            if key not in new:
-                self.mismatches.append(f"Missing from new: {full_path}")
-                continue
-            if key not in old:
-                self.mismatches.append(f"Extra in new: {full_path}")
-                continue
-
-            old_val = old[key]
-            new_val = new[key]
-            nested_diffs = self._nested_fields(known_diffs, key)
-            nested_unreliable = self._nested_fields(unreliable, key)
-
-            if nested_diffs or nested_unreliable:
-                child_diffs_path = _qualify(diffs_path, key)
-                if isinstance(old_val, list) and isinstance(new_val, list):
-                    if len(old_val) != len(new_val):
-                        self.mismatches.append(
-                            f"{full_path} count: old={len(old_val)}, new={len(new_val)}"
-                        )
-                    for i, (old_item, new_item) in enumerate(zip(old_val, new_val)):
-                        item_path = f"{full_path}[{i}]"
-                        if isinstance(old_item, Mapping) and isinstance(new_item, Mapping):
-                            self.compare(
-                                old_item,
-                                new_item,
-                                nested_diffs,
-                                item_path,
-                                child_diffs_path,
-                                unreliable=nested_unreliable,
-                            )
-                        elif old_item != new_item:
-                            self.mismatches.append(
-                                f"{item_path}: old={old_item!r}, new={new_item!r}"
-                            )
-                elif isinstance(old_val, Mapping) and isinstance(new_val, Mapping):
-                    self.compare(
-                        old_val,
-                        new_val,
-                        nested_diffs,
-                        full_path,
-                        child_diffs_path,
-                        unreliable=nested_unreliable,
-                    )
-                elif old_val != new_val:
-                    self.mismatches.append(f"{full_path}: old={old_val!r}, new={new_val!r}")
-            elif old_val != new_val:
-                self.mismatches.append(f"{full_path}: old={old_val!r}, new={new_val!r}")
